refactor(cmd): move version logic
