+++ cmd/root.go
@@ -1,8 +1,6 @@
 package cmd
 
 import (
-	"fmt"
-
 	"github.com/spf13/cobra"
 )
 
@@ -13,38 +11,10 @@ const banner = `         __   ___.
 |__|_|  /__|_ \|___  /__|   |__|   
       \/     \/    \/              `
 
-var (
-	version   string
-	buildTime string
-)
-
 var rootCmd = &cobra.Command{
 	Use:   "mkbrr",
 	Short: "A tool to inspect and create torrent files",
 	Long:  banner + "\n\nmkbrr is a tool to create and inspect torrent files.",
-}
-
-var versionCmd = &cobra.Command{
-	Use:   "version",
-	Short: "Print version information",
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Printf("Version: %s\n", version)
-		fmt.Printf("Build Time: %s\n", buildTime)
-	},
-	DisableFlagsInUseLine: true,
-}
-
-func SetVersion(v, bt string) {
-	version = v
-	buildTime = bt
-}
-
-func init() {
-	versionCmd.SetUsageTemplate(`Usage:
-  {{.CommandPath}}
-
-Prints the version and build time information for mkbrr.
-`)
 }
 
 func Execute() error {
