+++ main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"os"
+	"runtime/debug"
 
 	"github.com/autobrr/mkbrr/cmd"
 )
@@ -12,6 +13,12 @@ var (
 )
 
 func main() {
+	if version == "dev" {
+		if info, ok := debug.ReadBuildInfo(); ok {
+			version = info.Main.Version
+		}
+	}
+
 	cmd.SetVersion(version, buildTime)
 	if err := cmd.Execute(); err != nil {
 		os.Exit(1)
